Ciera compilation success
Fixed a couple places where a normal language would have easily worked
out that the math on the RHS should be done as type real and therefore
the transient variable on the LHS of the assignment should be declared
as real.  But, BP (maybe Ciera) is fussy about this stuff and bases the
type for the computation on the LHS of the expression.  So, y = x + 1.0;
must become y = 1.0 + x; when x is integer.